reservation_system/requirements.txt
Django==5.1.1
djangorestframework==3.15.2
drf-spectacular==0.27.2
python-docx==1.1.2
